DPDK: Refactor and move functions to dpdkutil
Creating the perf tests required hoisting some private functions
and their dependencies to a shared module, DpdkUtil.

- Move DpdkTestResources class to DpdkUtil
- Move and modify verify_dpdk_build and it's cousins to return
the DpdkTestResources instances for the test.
